requirements.txt
Django==5.0.6
djangorestframework==3.15.1
boto3==1.34.105
chromadb==0.5.0
langchain==0.1.20
django-cors-headers==4.3.1
openai==1.30.1